Module Sc_cbmc.Cbmc_data

Json interface for CBMC.

val only_data : ?pp_message:(Cbmc_types.DATA.message -> unit) -> ?pp_program_info:(string -> unit) -> 'a Cbmc_types.DATA.data list -> 'a list

Returns the data content of a list.

val uncovered_properties : labelized_file:[ `C | `labelized ] Sc_sys.File.TYPES.file -> variables:Cbmc_types.ENV.var_environment -> Cbmc_types.DATA.property list -> [ `simple ] Sc_lib.Types.Label.t list -> [ `simple ] Cbmc_types.DATA.analysis_env

Given a list of proof objectives and properties, returns the properties corresponding to unknown proof objectives. They are matched with their location in the C file.

module Make (Conf : Sc_lib.Config.T) (RuntimeDB : Sc_lib.Tool.RuntimeDatabase) : sig ... end

Defines the interface for calling CBMC.