Module Cbmc_data.Make

Defines the interface for calling CBMC.

Parameters

module Conf : Sc_lib.Config.T

module RuntimeDB : Sc_lib.Tool.RuntimeDatabase

Signature

val cbmc_get_properties : rdb:RuntimeDB.runtime_data -> working_dir:Sc_sys.File.TYPES.dir -> funname:string -> files:string list -> Cbmc_types.OPTIONS.cbmc_options -> Cbmc_types.DATA.property list Cbmc_types.DATA.data list Lwt.t

Starts a CBMC process that reads the file and return the list of properties

val cbmc_get_cover_objectives : rdb:RuntimeDB.runtime_data -> working_dir:Sc_sys.File.TYPES.dir -> funname:string -> files:string list -> Cbmc_types.OPTIONS.cbmc_options -> Cbmc_types.DATA.property list Cbmc_types.DATA.data list Lwt.t

Starts a CBMC process that reads the file and return the list of cover objectives (i.e., instructions to cover)

val cbmc_get_clabels : rdb:RuntimeDB.runtime_data -> lbls:[ `simple ] Sc_lib.Types.Label.t list -> working_dir:Sc_sys.File.TYPES.dir -> funname:string -> files:string list -> Cbmc_types.OPTIONS.cbmc_options -> Cbmc_types.DATA.label list Cbmc_types.DATA.data list Lwt.t

Starts a CBMC process that reads the file and return the list of labels

val cbmc_cover_analysis : rdb:RuntimeDB.runtime_data -> working_dir:Sc_sys.File.TYPES.dir -> funname:string -> files:string list -> to_cover:[ `simple ] Cbmc_types.DATA.analysis_env -> Cbmc_types.OPTIONS.cbmc_options -> Cbmc_types.DATA.cbmc_cover_output Cbmc_types.DATA.data list Lwt.t

Starts a CBMC process that analyzes the file and returns the test cases covering the labels.

val cbmc_assert_analysis : rdb:RuntimeDB.runtime_data -> working_dir:Sc_sys.File.TYPES.dir -> funname:string -> files:string list -> to_cover:[ `simple ] Cbmc_types.DATA.analysis_env -> Cbmc_types.OPTIONS.cbmc_options -> Cbmc_types.DATA.cbmc_assert_output Cbmc_types.DATA.data list Lwt.t

Starts a CBMC process that checks if the labels are uncoverable by asserting the negation of their condition.

val cbmc_clabel_analysis : rdb:RuntimeDB.runtime_data -> working_dir:Sc_sys.File.TYPES.dir -> funname:string -> files:string list -> to_cover:[ `simple ] Cbmc_types.DATA.analysis_env -> Cbmc_types.OPTIONS.cbmc_options -> Cbmc_types.DATA.cbmc_assert_output Cbmc_types.DATA.data list Lwt.t

Same as cbmc_assert_analysis, but uses the error labels feature of CBMC instead of assertions.

val goals_to_test_cases : ?pp_message:(Cbmc_types.DATA.message -> unit) -> ?pp_program_info:(string -> unit) -> [ `simple ] Cbmc_types.DATA.analysis_env -> Cbmc_types.DATA.cbmc_cover_output Cbmc_types.DATA.data list -> Cbmc_types.DATA.cbmc_result

Returns the list of inputs and their associated C value.

val assert_data_list_to_test_cases : ?pp_message:(Cbmc_types.DATA.message -> unit) -> ?pp_program_info:(string -> unit) -> [ `simple ] Cbmc_types.DATA.analysis_env -> Cbmc_types.DATA.cbmc_assert_output Cbmc_types.DATA.data list -> Cbmc_types.DATA.cbmc_result

Returns the list of inputs and their associated C value.