Module Cbmc_types.DATA

type json_options = {

oarguments : string list;
(*
The list of files to analyze.
*)

ofunction : string;
(*
Sets the main function name.
*)

ounwind : int option;
(*
Sets the max number of unwindings. None = infinite unwind
*)

oproperties : string list option;
(*
Restricts the properties to check. None = all properties.
*)

oshow_properties : bool;
(*
Show the properties, but don't run the analysis.
*)

ocover : string option;
(*
Uses a specific coverage criterion. In our case we only use the "cover" coverage to aim at specific properties to cover.
*)

oerror_label : string list option;
(*
Checks that labels are uncoverable.
*)

opointer_check : bool;
(*
Enables pointer checks.
*)

}

The options for the CBMC tool. They will be passed as a JSON value to the tool with the option `--json-interface`. https://diffblue.github.io/cbmc//man/cbmc.html

Output of the tool. They are received as JSON values. All the following types are elements structuring CBMC's answer. These types should be abstract, except we need them for testing.

type source_location = {

slfile : string;

slfunction : string option;

slline : string;

slworking_dir : string option;

}

A location in the project.

type message = {

mtext : string;

mtyp : string;

msource_location : source_location option;

}

A custom message from CBMC.

type status =

| Success

| Failure_

| Unknown of string

type 'a data =

| ProgramInfo of string

| Message of message

| Data of 'a

| CProverStatus of status

type property_class =

| ErrorLabel

| Assertion

| Coverage

| Unknown of string

type property = {

pclass : property_class;

pdescription : string;

pexpression : string;

pname : string;

psource_location : source_location;

}

type label = property

type base_value =

| Value of {

vbinary : string option;

vdata : string;

vname : string;

vtype : string option;

vwidth : int option;

}

| Unknown

type structure_field = {

sfname : string;

sfvalue : value;

}

and structured = {

smembers : structure_field list;

sname : string;

}

and value =

| Base of base_value

| Structured of structured

| Array of (int * value) list

| Union of structure_field

type input = {

iid : string;

ivalue : value;

}

Output of cover analysis.

type test = {

tcovered_goals : string list;

tinputs : input list;

}

type goal = {

gdescription : string;

ggoal : string;

gsourceLocation : source_location;

gstatus : status;

}

type goals_details = {

gdgoals : goal list;

gdgoals_covered : int;

gdtotal_goals : int;

}

type cbmc_cover_output =

| Goals of goals_details

| Tests of test list

Output of assert analysis

type function_ = {

fdisplay_name : string;

fidentifier : string;

fsource_location : source_location;

}

type location_only = {

lohidden : bool;

losource_location : source_location;

lostep_type : string;

lothread : int;

}

type function_info = {

fifunction : function_;

fihidden : bool;

fiinternal : bool;

fistep_type : string;

fithread : int;

fisource_location : source_location option;

}

type assignment = {

atype : string;

ahidden : bool;

ainternal : bool;

alhs : string;

amode : string;

asource_location : source_location;

astep_type : string;

athread : int;

avalue : value;

areason : string option;

}

type input_step = {

ihidden : bool;

iinputID : string;

iinternal : bool;

imode : string;

isource_location : source_location;

istep_type : string;

ithread : int;

ivalues : value list;

}

type failure_step = {

fshidden : bool;

fsinternal : bool;

fsproperty : string;

fsreason : string;

fssource_location : source_location;

fsstep_type : string;

fsthread : int;

}

type output_step = {

ohidden : bool;

ointernal : bool;

ostep_type : string;

othread : int;

omode : string;

oid : string;

oloc : source_location;

ovalues : value list;

}

type instruction =

| Function of function_info

| Assignment of assignment

| Location of location_only

| FailureStep of failure_step

| Input of input_step

| Output of output_step

type assertion_check = {

acdescription : string;

acproperty : string;

acsource_location : source_location option;

acstatus : status;

actrace : instruction list option;

}

type cbmc_assert_output = assertion_check list

type cbmc_result

The CBMC's results.

val empty_cbmc_result : cbmc_result

val add_tests : (Sc_lib.Ctypes_bridge.literal_binding * Basics.Ints.t) list -> cbmc_result -> cbmc_result

add_tests tests covered res Adds `tests` covering the properties whose identifiers are in `covered`. If the new tests do not cover new labels, the lazy suspension is not even forced and tests are not added to the result. If the new tests cover more labels and they are strictly less than the one already registered, replaces the tests in the result. Otherwise, adds the new tests and adds the covered labels.

val add_non_valid_extra_property : assertion_check -> cbmc_result -> cbmc_result

Adds an assertion check as a non valid extra property (non valid = not proven, but not necessarily invalid) x

val add_uncoverable : Sc_lib.Types.Label.any Sc_lib.Types.Label.t -> cbmc_result -> cbmc_result

Adds a label as uncoverable.

val get_tests : cbmc_result -> (Sc_lib.Ctypes_bridge.literal_binding * Basics.Ints.t) list

Returns the tests registered.

val get_covered : cbmc_result -> Basics.Ints.t

Returns the covered labels.

val get_uncoverable : cbmc_result -> Basics.Ints.t

Returns the uncoverable labels. If there are non valid extra properties, returns the empty set.

module PropertyMap : sig ... end

A map binding properties (from their name).

type simple_label_env = {

proof_objectives : [ `simple ] Sc_lib.Types.Label.t PropertyMap.t;

extra_required_properties : label list;

variables : ENV.var_environment;

}

type 'a analysis_env =

| SimpleLabelEnv : simple_label_env -> [ `simple ] analysis_env

The information required for starting CBMC.

module Encoding : sig ... end

Most of these endodings are exported for testing.