Module Cbmc_types.ENV

type size_detail =

| AsFormal of {

var : string;

max : int;

}
(*
Size is represented as a formal.
*)

| AsConst of int
(*
Size is known
*)

type mem_details =

| NotAPointer
(*
A simple value
*)

| Pointer of {

first_size : size_detail;

sub_pointers : size_detail list;

}
(*
A pointer. Its size is specified by the `first_size` field while the `sub_pointers` represent the sub_arrays if the pointer is a multi dimensional array.
*)

The size of an array/string. It can either be represented as a variable or a known constant.

type origin =

| Global of {

orig_name : string;

harness_arg_name : string;

}

| Formal of string

| Internal

The representation of a harness formal. Harness formals can either initialize globals, be an original formal or a fresh formal (for array size for example).

type var_environment = (Sc_lib.Cil_utils.var * mem_details * origin) Basics.StrMap.t

A map binding a variable name (as appearing in the harness) to its size and some information.