Module Printer.LiteralMemory

Literal representation of C typed memory

type t = assigns * LiteralHeap.t

A C memory consists of initialized typed variables, along with a literal heap.

and assigns

type c_code = {

pp_heap : static:bool -> Basics.PPrt.pu;
(*
same as LiteralHeap.print
*)

pp_globals : Basics.PPrt.pu;
(*
prints assignment of already declared variables
*)

pp_locals : Basics.PPrt.pu;
(*
prints declarations and initializations for local variables
*)

}

C code used to declare and initialize C memory

val as_c_code : globals:Sc_lib.Cil_utils.vars -> locals:Sc_lib.Cil_utils.vars -> t -> c_code