Module Sc_cbmc.Harness

type t

The harness representation.

val oracle_property_identifier : string

The property name of the oracle assertion used for fixtures.

val generate : project:_ Sc_project.Types.project -> target:[> `C ] Sc_sys.File.t -> cbmc_driver:[> `H ] Sc_sys.File.t -> t

generate ~dry ~project ~target ~cbmc_driver

Generates a harness to pass to CBMC and writes it in target. The input format is as follows: simple variables: "varname" arrays & pointers: "arr0, arr1, ..., arrsize-1" structures & unions: "t.field" cbmc_driver is the file used as a driver. It depends on the selected CBMC mode.

If dry is true, only initializes the internal representation of the harness.

val entrypoint : t -> string

val test_to_literal : t -> Types.DATA.test -> Sc_values.literal_binding

Reads a test generated by CBMC and returns its literal binding representation.

val trace_to_literal : t -> Types.DATA.instruction list -> Sc_values.literal_binding

Reads a trace generated by a CBMC counter example and returns its literal binding representation.

val num_symbolic_variables : t -> int

Returns the number of symbolic variables used in the harness.