Module Sc_cbmc.Runner

Helpers to run CBMC.

type 'a cbmc_run = store:Sc_store.t -> runner_options:Types.runner_options -> entrypoint:string -> files:[ `C ] Sc_sys.File.t list -> harness:Harness.t -> Types.OPTIONS.t -> ('a Lwt_stream.t * (unit -> unit Lwt.t)) Lwt.t

val uncovered_properties : mode:Types.OPTIONS.mode -> harness_file:[ `C ] Sc_sys.File.TYPES.file -> labelized_file:[ `C | `labelized ] Sc_sys.File.TYPES.file -> cbmc_props:Types.DATA.property list -> already_decided:Basics.Ints.t -> labels:Sc_C.Cov_label.simple list -> entrypoint:string -> [ `simple ] Types.analysis_env

Given a list of proof objectives and properties, returns the properties corresponding to unknown proof objectives. They are matched with their location in the C file.

val cbmc_get_properties : Types.DATA.property list Types.DATA.cell cbmc_run

Starts a CBMC process that reads the file and return the list of properties

val cbmc_get_cover_objectives : Types.DATA.property list Types.DATA.cell cbmc_run

Starts a CBMC process that reads the file and return the list of cover objectives (i.e., instructions to cover)

val cbmc_get_clabels : lbls:Sc_C.Cov_label.simple list -> Types.DATA.property list Types.DATA.cell cbmc_run

Starts a CBMC process that reads the file and return the list of labels

val cbmc_cover_analysis : to_cover:[ `simple ] Types.analysis_env -> Types.DATA.cbmc_cover_output Types.DATA.cell cbmc_run

Starts a CBMC process that analyzes the file and returns the test cases covering the labels.

val cbmc_assert_analysis : to_cover:[ `simple ] Types.analysis_env -> Types.DATA.cbmc_assert_output Types.DATA.cell cbmc_run

Starts a CBMC process that checks if the labels are uncoverable by asserting the negation of their condition.

val cbmc_clabel_analysis : to_cover:[ `simple ] Types.analysis_env -> Types.DATA.cbmc_assert_output Types.DATA.cell cbmc_run

Same as cbmc_assert_analysis, but uses the error labels feature of CBMC instead of assertions.